requirements.txt
requests>=2.28.0
pyyaml>=6.0
pandas>=1.5.0
python-dotenv>=0.21.0
schedule>=1.2.0
aiohttp>=3.8.0